requirements.txt
python-telegram-bot==20.7
pydantic==2.5.3
pyyaml==6.0.1

